chore: re-structure files, move list to its own component
