+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Router } from '@reach/router'
 import RestrictedRoute from './views/restricted'
-import Header from './components/Header/Header'
-import Footer from './components/Footer/Footer'
+import Header from './components/Header'
+import Footer from './components/Footer'
 import styled from '@emotion/styled'
 const Login = React.lazy(() => import(`./views/Login/Login`))
 const Start = React.lazy(() => import(`./views/Start/Start`))
