+++ src/components/QR.tsx
