+++ src/components/__tests__/ButtonLink.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import ButtonLink from '../ButtonLink'
-import { render } from '../../../utils/test-utils'
+import { render } from '../../utils/test-utils'
 
 describe('components/ButtonLink', () => {
   it('renders a ButtonLink', async () => {
