+++ src/components/__tests__/Footer.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Footer from '../Footer'
-import { render } from '../../../utils/test-utils'
+import { render } from '../../utils/test-utils'
 
 describe('components/Footer', () => {
   it('renders Footer component', async () => {
