+++ src/components/__tests__/Header.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Header from '../Header'
-import { render } from '../../../utils/test-utils'
+import { render } from '../../utils/test-utils'
 
 describe('components/Header', () => {
   it('renders Header component', async () => {
