+++ src/components/__tests__/Input.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import Input from '../Input'
-import { render } from '../../../utils/test-utils'
+import { render } from '../../utils/test-utils'
 
 describe('components/Input', () => {
   it('renders an Input', async () => {
