+++ src/components/__tests__/List.spec.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react'
+import List from '../List'
+import { render } from '../../utils/test-utils'
+
+describe('components/List', () => {
+  it('renders a List', async () => {
+    const { container } = render(<List />)
+
+    expect(container).toMatchSnapshot()
+  })
+})
